% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_longest_trip.R
\name{add_trip_middle}
\alias{add_trip_middle}
\title{Add row with location of middle of trip}
\usage{
add_trip_middle(.longest_trip)
}
\arguments{
\item{.longest_trip}{(tbl_df) Tibble containing information
about the longest trip}
}
\value{
(tbl_df) Longest trip data with a 3rd row added
}
\description{
Used to add pirate ship in trip path
}
