% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_ships_data.R
\name{assert_ships_data}
\alias{assert_ships_data}
\title{Assert cleaned ships data is as expected}
\usage{
assert_ships_data(.ships_data)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}
}
\value{
(logical | error) TRUE if passed, error otherwise
}
\description{
Assert cleaned ships data is as expected
}
