% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_ships_data_raw.R
\name{assert_ships_data_raw}
\alias{assert_ships_data_raw}
\title{Assert raw ships data is as expected}
\usage{
assert_ships_data_raw(.ships_data)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}
}
\value{
(logical | error) TRUE if passed, error otherwise
}
\description{
Assert raw ships data is as expected
}
