% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tbl_general.R
\name{assert_tbl_col_names}
\alias{assert_tbl_col_names}
\title{Assert that a tibble has expected column names}
\usage{
assert_tbl_col_names(.tbl, .col_spec)
}
\arguments{
\item{.tbl}{(tbl_df) Tibble}

\item{.col_spec}{(tbl_df) Column spec tibble}
}
\value{
(tbl_df) Returns input if passed, error otherwise
}
\description{
Assert that a tibble has expected column names
}
