% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tbl_general.R
\name{assert_tbl_no_duplicate_rows}
\alias{assert_tbl_no_duplicate_rows}
\title{Assert that a tibble has no duplicate rows}
\usage{
assert_tbl_no_duplicate_rows(.tbl)
}
\arguments{
\item{.tbl}{(tbl_df) Tibble}
}
\value{
(tbl_df) Returns input if passed, error otherwise
}
\description{
Assert that a tibble has no duplicate rows
}
