% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tbl_general.R
\name{assert_tbl_no_missing_values}
\alias{assert_tbl_no_missing_values}
\title{Assert that a tibble has no missing values}
\usage{
assert_tbl_no_missing_values(.tbl)
}
\arguments{
\item{.tbl}{(tbl_df) Tibble}
}
\value{
(tbl_df) Returns input if passed, error otherwise
}
\description{
Assert that a tibble has no missing values
}
