% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tbl_general.R
\name{assert_tbl_tibble}
\alias{assert_tbl_tibble}
\title{Assert that input is a tibble}
\usage{
assert_tbl_tibble(.tbl)
}
\arguments{
\item{.tbl}{(tbl_df) Tibble}
}
\value{
(tbl_df) Returns input if passed, error otherwise
}
\description{
Assert that input is a tibble
}
