% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_longest_trip.R
\name{calc_longest_trip}
\alias{calc_longest_trip}
\title{Calculate the longest trip for a given ship}
\usage{
calc_longest_trip(.ships_data)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data
containing data from a single ship}
}
\value{
(tbl_df) Tibble containing information
  about the longest trip
}
\description{
Calculate the longest trip for a given ship
}
