% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ships_data.R
\name{clean_remove_duplicate_rows}
\alias{clean_remove_duplicate_rows}
\title{Remove duplicate rows}
\usage{
clean_remove_duplicate_rows(.ships_data)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}
}
\value{
(tbl_df) Ships data without duplicate rows
}
\description{
Remove duplicate rows
}
