% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ships_data.R
\name{clean_rename_cols}
\alias{clean_rename_cols}
\title{Rename columns of ships data}
\usage{
clean_rename_cols(.ships_data)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}
}
\value{
(tbl_df) Ships data with renamed columns
}
\description{
Rename columns of ships data
}
