% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ships_data.R
\name{clean_ships_data}
\alias{clean_ships_data}
\title{Clean ships data}
\usage{
clean_ships_data(.ships_data)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}
}
\value{
(tbl_df) Cleaned ships data
}
\description{
Clean ships data
}
