% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ship_name.R
\name{filter_ship_name}
\alias{filter_ship_name}
\title{Filter ships data by ship name}
\usage{
filter_ship_name(.ships_data, .ship_name)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}

\item{.ship_name}{(character) Ship name}
}
\value{
(tbl_df) Ships data filtered by ship name
}
\description{
Filter ships data by ship name
}
