% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ship_type.R
\name{filter_ship_type}
\alias{filter_ship_type}
\title{Filter ships data by ship type}
\usage{
filter_ship_type(.ships_data, .ship_type)
}
\arguments{
\item{.ships_data}{(tbl_df) Ships data}

\item{.ship_type}{(character) Ship type}
}
\value{
(tbl_df) Ships data filtered by ship type
}
\description{
Filter ships data by ship type
}
