% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trip_note.R
\name{gen_trip_note}
\alias{gen_trip_note}
\title{Generate text for the longest trip note}
\usage{
gen_trip_note(.longest_trip)
}
\arguments{
\item{.longest_trip}{(tbl_df) Tibble containing
information about the trip}
}
\value{
(character) Longest trip note
}
\description{
Generate text for the longest trip note
}
