% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tbl_general.R
\name{not_missing}
\alias{not_missing}
\title{Check if all values of an object are not missing}
\usage{
not_missing(.x)
}
\arguments{
\item{.x}{(atomic | list) Vector or list}
}
\value{
(logical) Whether any values are missing
}
\description{
Check if all values of an object are not missing
}
