% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trip.R
\name{plot_trip}
\alias{plot_trip}
\title{Plot a map with the start, middle and end of a trip}
\usage{
plot_trip(.trip)
}
\arguments{
\item{.trip}{(tbl_df) Tibble containing information
about the trip}
}
\value{
(leaflet) Leaflet htmlwidget map
}
\description{
Plot a map with the start, middle and end of a trip
}
